Inventory-Backend-FastAPI/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
email-validator==2.1.0.post1
python-multipart==0.0.9
httpx==0.26.0
alembic==1.13.1
